requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
camelot-fork==0.20.1
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.0
et-xmlfile==1.1.0
ghostscript==0.7
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.4
mdurl==0.1.2
numpy==1.26.3
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==23.2
pandas==1.5.3
pdfminer.six==20221105
pillow==10.2.0
protobuf==4.25.2
pyarrow==15.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
pypdf==3.17.4
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.32.1
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
six==1.16.0
smmap==5.0.1
streamlit==1.30.0
tabulate==0.9.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
XlsxWriter==3.1.9
zipp==3.17.0
